PHP 8.1 | Indexable::save() prevent null to non-nullable deprecation notice
The `$primary_focus_keyword` property is a (dynamic/managed via magic methods) property on the `ORM` class. This property may or may not be set at the time of a save.

As of PHP 8.1, passing `null` to a non-nullable PHP native function will generate a deprecation notice, which is what will happen when the `$primary_focus_keyword` property has not been set (yet) and is subsequently passed to the PHP native `strlen()` function.

Verifying that the property is a string before passing it to `strlen()` prevents the notice without BC-break.

Note: I've elected to verify the usability of the property using the relatively strict `is_string()` as the property is _documented_ as expecting to be a string.
This does mean that the `$primary_focus_keyword` can now no longer be passed a `Stringable` object, nor be a non-string scalar.

IMO, for non-string scalars, checking the string length and using `mb_substr()` on the value doesn't really make much sense anyway.
And as these classes are only intended for "internal" use and the Yoast plugins do not use `Stringable` objects for the primary focus keyword, supporting those was deemed unnecessary.
If this would at a later point in time become something which should be supported, the condition can then be updated to also check for `Stringable` objects.

Includes updating existing unit tests to expect the extra call to the `get()` method to retrieve the property.

Refs:
* https://wiki.php.net/rfc/deprecate_null_to_scalar_internal_arg